updating gridconfig generator file
