+++ live-editing/configs/GridConfigGenerator.ts
@@ -457,12 +457,13 @@ export class GridConfigGenerator implements IConfigGenerator {
             component: FinJSDemoComponent,
             additionalFiles: ["/src/app/grid/services/localData.service.ts", "/src/app/grid/services/financialData.ts"],
             appModuleConfig: new AppModuleConfig({
-                imports: [ IgxGridModule, IgxButtonGroupModule, IgxIconModule, IgxSliderModule,
-                    IgxSwitchModule, IgxRippleModule, FinJSDemoComponent, LocalDataService],
+                imports: [ IgxGridModule, IgxButtonGroupModule, IgxIconModule, IgxSliderModule, IgxToggleModule,
+                    IgxButtonModule, IgxExcelExporterService, IgxSwitchModule, IgxRippleModule, FinJSDemoComponent,
+                    LocalDataService],
                 ngDeclarations: [FinJSDemoComponent],
-                ngImports: [IgxGridModule, IgxButtonGroupModule, IgxIconModule, IgxSliderModule,
-                    IgxSwitchModule, IgxRippleModule, IgxGridModule.forRoot()],
-                ngProviders: [LocalDataService]
+                ngImports: [IgxGridModule, IgxButtonGroupModule, IgxIconModule, IgxSliderModule, IgxToggleModule,
+                    IgxButtonModule, IgxSwitchModule, IgxRippleModule, IgxGridModule.forRoot()],
+                ngProviders: [LocalDataService, IgxExcelExporterService]
             })
         }));
