--- src/app/grid/services/mockupData.ts
@@ -1,74 +1,0 @@
-const MOCKFINANCEDATA: any[] = [
-    {
-      "IndGrou": "Airlines",
-      "IndSect": "Consumer, Cyclical",
-      "IndSubg": "Airlines",
-      "SecType": "PUBLIC",
-      // tslint:disable-next-line:object-literal-sort-keys
-      "CpnTyp": "FIXED",
-      "IssuerN": "AMERICAN AIRLINES GROUP",
-      "Moodys": "WR",
-      "Fitch": "N.A.",
-      "DBRS": "N.A.",
-      "CollatT": "NEW MONEY",
-      "Curncy": "USD",
-      "Security": "001765866 Pfd",
-      "sector": "Pfd",
-      "CUSIP": "1765866",
-      "Ticker": "AAL",
-      "Cpn": "7.875",
-      "Maturity": "7/13/1939",
-      "KRD_3YR": 0.00006,
-      "RISK_COUNTRY": "",
-      "MUNI_SECTOR": "",
-      "ZV_SPREAD": 28.302,
-      "KRD_5YR": 0,
-      "KRD_1YR": -0.00187,
-      "PD_WALA": null,
-      "STRUCTURE": "",
-      "PRIN_BE_CUM_LOSS": 0,
-      "PD_CPR_3M": 100,
-      "KTRD_2YR": 0.00137,
-      "TICKER": "NMOTR 2016-A A1",
-      "COLLATERAL_TYPE": "",
-      "PRIN_LOSS_DOLLAR": "0",
-      "KRD_2YR": 0.00136,
-      "PUT_CALL": "",
-      "PD_CPR_WALEQ": 24.3,
-      "Amt Out": "6000000",
-      "Amt Issu": "6000000",
-      "Curr Net": 3760,
-      "Cpn Freq": 4,
-      "Bid Modi": "",
-      "OAS DV01": 19,
-      "Mid Conv": "",
-      "Net P&L": 0,
-      "MTD Rlzd": "",
-      "MTD Unrl": "",
-      "YTD Rlzd": "",
-      "YTD Unrl": "",
-      "Amort P&": null,
-      "Acc Int": null,
-      "Sec Typ": "Preferred Stock",
-      "Yest MTM": "",
-      "Mid Px": "11.5",
-      "YAS Bond": "",
-      "YAS OAS": "",
-      "Imp Px S": "",
-      "DDM Inte": "",
-      "Mid Curr": "",
-      "TDS Cred": 0,
-      "D to Mat": 5506,
-      "Dur 1": 5.294,
-      "WAM": "",
-      "WA Life": "",
-      "Typ": "",
-      "Tranche": "",
-      "Factor D": "",
-      "ECON_YIELD": 1.62911,
-      "MKT_NOTION": "$47,338,486",
-      "Background Mkt Notion Alloc%": "0.0517838721652916000000000%",
-      "Weighted Yield by MKT Notion": "0.0843616239831982000000000%"
-    }];
-
-export { MOCKFINANCEDATA };
